Config.prod_dir is now configurable a la working_directory
this allows to specify a rack folder that doesnt necessarily live inside of the packed smtag library.

this is convenient for the cases in which the trained moldes are too big (eg densenet) and cannot be commited to the git repository.

in this scenarios the user of py-smtag would need to manually download the models and point to them with the new env variable `SMTAG_PRODUCTION_DIRECTORY` or by using the new CLI flags "--production_directory and its shorted version -p